chore: move simple example to main.dart so it will be displayed on pub.dev
