+++ example/lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:mobile_scanner/mobile_scanner.dart';
 import 'package:mobile_scanner_example/screens/mobile_scanner_advanced.dart';
-import 'package:mobile_scanner_example/screens/mobile_scanner_simple.dart';
 
 void main() {
   runApp(
@@ -8,6 +8,71 @@ void main() {
   );
 }
 
+/// Implementation of Mobile Scanner example with simple configuration
+class MobileScannerSimple extends StatefulWidget {
+  /// Constructor for simple Mobile Scanner example
+  const MobileScannerSimple({super.key});
+
+  @override
+  State<MobileScannerSimple> createState() => _MobileScannerSimpleState();
+}
+
+class _MobileScannerSimpleState extends State<MobileScannerSimple> {
+  Barcode? _barcode;
+
+  Widget _barcodePreview(Barcode? value) {
+    if (value == null) {
+      return const Text(
+        'Scan something!',
+        overflow: TextOverflow.fade,
+        style: TextStyle(color: Colors.white),
+      );
+    }
+
+    return Text(
+      value.displayValue ?? 'No display value.',
+      overflow: TextOverflow.fade,
+      style: const TextStyle(color: Colors.white),
+    );
+  }
+
+  void _handleBarcode(BarcodeCapture barcodes) {
+    if (mounted) {
+      setState(() {
+        _barcode = barcodes.barcodes.firstOrNull;
+      });
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: const Text('Simple Mobile Scanner')),
+      backgroundColor: Colors.black,
+      body: Stack(
+        children: [
+          MobileScanner(onDetect: _handleBarcode),
+          Align(
+            alignment: Alignment.bottomCenter,
+            child: Container(
+              alignment: Alignment.bottomCenter,
+              height: 100,
+              color: const Color.fromRGBO(0, 0, 0, 0.4),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                children: [
+                  Expanded(child: Center(child: _barcodePreview(_barcode))),
+                ],
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
+
+/// Homepage for example app with selection between basic and advanced screen.
 class _ExampleHome extends StatelessWidget {
   const _ExampleHome();
