--- example/lib/screens/mobile_scanner_simple.dart
@@ -1,66 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:mobile_scanner/mobile_scanner.dart';
-
-/// Implementation of Mobile Scanner example with simple configuration
-class MobileScannerSimple extends StatefulWidget {
-  /// Constructor for simple Mobile Scanner example
-  const MobileScannerSimple({super.key});
-
-  @override
-  State<MobileScannerSimple> createState() => _MobileScannerSimpleState();
-}
-
-class _MobileScannerSimpleState extends State<MobileScannerSimple> {
-  Barcode? _barcode;
-
-  Widget _buildBarcode(Barcode? value) {
-    if (value == null) {
-      return const Text(
-        'Scan something!',
-        overflow: TextOverflow.fade,
-        style: TextStyle(color: Colors.white),
-      );
-    }
-
-    return Text(
-      value.displayValue ?? 'No display value.',
-      overflow: TextOverflow.fade,
-      style: const TextStyle(color: Colors.white),
-    );
-  }
-
-  void _handleBarcode(BarcodeCapture barcodes) {
-    if (mounted) {
-      setState(() {
-        _barcode = barcodes.barcodes.firstOrNull;
-      });
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(title: const Text('Simple Mobile Scanner')),
-      backgroundColor: Colors.black,
-      body: Stack(
-        children: [
-          MobileScanner(onDetect: _handleBarcode),
-          Align(
-            alignment: Alignment.bottomCenter,
-            child: Container(
-              alignment: Alignment.bottomCenter,
-              height: 100,
-              color: const Color.fromRGBO(0, 0, 0, 0.4),
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                children: [
-                  Expanded(child: Center(child: _buildBarcode(_barcode))),
-                ],
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
